requirements.txt
opencv-python<4.10
numpy<2
mediapipe
python-dotenv
